bridge/requirements.txt
opencv-python==4.8.1.78
requests==2.31.0
numpy==1.24.3
Pillow==10.1.0
configparser==7.0.0